requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
supabase==2.0.0
PyPDF2==3.0.1
PyMuPDF==1.23.8
requests==2.31.0
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.8.1.78
